feat: initiate nix (#14)
## Purpose

**As an engineer**
**I want to** shift my configuration to utilize Nix
**So that** I can declaratively standup my workstation environment
without the need for custom bash scripts and managing their order.

**Issue link (if applicable):** #13

## Implementation Notes

* Shift to Nix for most configuration
* GUI apps appear to be better handled by homebrew so they appear in
`/Applications` instead of `~/Applications`. This influences
applications not showing up Launchpad and the Finder applications view
which is a worse DevX
* Pivoting to utilize nix-darwin, will allow configuring Homebrew via
Nix as well as allow setting Mac configuration in support of #8.
* Currently the old approach for homebrew is living alongside nix.
Utilizing Nix Darwin will be a future iteration.
* Docs will be updated at the end of the overhaul. This is an iterative
PR into a larger feature branch.